Refactor Redux filters by merging action files, reducers, etc.
