+++ client/src/actions/filters.js
@@ -1,0 +1,11 @@
+// SET_POINTS_FILTER
+export const setPointsFilter = (text) => ({
+  type: 'SET_POINTS_FILTER',
+  text
+});
+
+// SET_TOPICS_FILTER
+export const setTopicsFilter = (text) => ({
+  type: 'SET_TOPICS_FILTER',
+  text
+});
