--- client/src/actions/pointsFilter.js
@@ -1,5 +1,0 @@
-// SET_POINTS_FILTER
-export const setPointsFilter = (text) => ({
-  type: 'SET_POINTS_FILTER',
-  text
-});
