--- client/src/actions/topicsFilter.js
@@ -1,5 +1,0 @@
-// SET_TOPICS_FILTER
-export const setTopicsFilter = (text) => ({
-  type: 'SET_TOPICS_FILTER',
-  text
-});
