--- client/src/reducers/pointsFilter.js
@@ -1,10 +1,0 @@
-const initialState = '';
-
-export default (prevState = initialState, action) => {
-  switch (action.type) {
-    case 'SET_POINTS_FILTER':
-      return action.text;
-    default:
-      return prevState;
-  }
-};
