+++ client/src/store/configureStore.js
@@ -1,13 +1,11 @@
 import { createStore, combineReducers } from 'redux';
 import topicsReducer from '../reducers/topics';
-import topicsFilterReducer from '../reducers/topicsFilter';
-import pointsFilterReducer from '../reducers/pointsFilter';
+import filtersReducer from '../reducers/filters';
 
 export default () => createStore(
   combineReducers({
     topics: topicsReducer,
-    topicsFilter: topicsFilterReducer,
-    pointsFilter: pointsFilterReducer
+    filters: filtersReducer
   }),
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
 );
